Wizard/Packages: Fix package recommendations tracking
The original tracking didn't work, because the `Select` didn't trigger `onOpenChange`, only on clicking the dropdown button.

All the analytics tracking was moved, so the Packages component is cleaner and the changes in the isOpen can be tracked via state setters.

Also decreased prop-drilling.
